Books: move access rights and templatex => ES6
