+++ book/static/js/es6_modules/books/accessrights/templates.js
@@ -1,26 +1,5 @@
-/**
- * @file Book access rights templates.
- * @copyright This file is part of <a href='http://www.fiduswriter.org'>Fidus Writer</a>.
- *
- * Copyright (C) 2013 Takuto Kojima, Johannes Wilm.
- *
- * @license This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <a href='http://www.gnu.org/licenses'>http://www.gnu.org/licenses</a>.
- *
- */
-
 /** A template for the book access rights overview */
-var tmp_book_access_right_overview = _.template('\
+export let bookAccessRightOverviewTemplate = _.template('\
     <div id="access-rights-dialog" title="<%- dialogHeader %>">\
         <div id="my-contacts" class="fw-ar-container">\
             <h3 class="fw-green-title">' + gettext("My contacts") + '</h3>\
@@ -48,9 +27,12 @@ var tmp_book_access_right_overview = _.template('\
                 <tbody class="fw-document-table-body fw-small"><%= collaborators %></tbody>\
             </table>\
         </div>\
-    </div>');
+    </div>'
+)
+
 /** A template for the book collaboration pane */
-var tmp_book_collaborators = _.template('<% _.each(collaborators, function(collaborator) { %>\
+export let bookCollaboratorsTemplate = _.template('\
+    <% _.each(collaborators, function(collaborator) { %>\
         <tr id="collaborator-<%- collaborator.user_id %>" data-id="<%- collaborator.user_id %>"\
         class="collaborator-tr <%- collaborator.rights %>" data-right="<%- collaborator.rights %>">\
             <td width="212">\
@@ -83,4 +65,5 @@ var tmp_book_collaborators = _.template('<% _.each(collaborators, function(colla
                 </span>\
             </td>\
         </tr>\
-    <% }) %>');
+    <% }) %>'
+)
