+++ book/templates/book/index.html
@@ -20,8 +20,6 @@
         <script type="text/javascript" src="{{ STATIC_URL }}js/document/stylelist.js"></script>
         <!-- jQuery Dropkick -->
         <script type="text/javascript" src="{{ STATIC_URL }}js/libs/jquery.dropkick-1.1.0.js"></script>
-        <script type="text/javascript" src="{{ STATIC_URL }}js/book-template-dialogs.js"></script>
-        <script type="text/javascript" src="{{ STATIC_URL }}js/bookaccessrights-helpers.js"></script>
         <script type="text/javascript" src="{{ STATIC_URL }}js/libs/jquery.dataTables.min.js"></script>
         <script type="text/javascript" src="{{ STATIC_URL }}js/document-helpers.js"></script>
         <!-- File export functions -->
